large restructuring of code and draggable pieces
Restructured a lot of things so that there are children and grandchildren components that can interact.

Made the board resizable and still need to add this to the squares. Right now the labels are what determine the size of the board since they stretch to 10% of the game.

Made pieces draggable, but still need to incorporate that in the move.

The white king still cannot capture for some reason.

Still no en passante or castling.